requirements.txt
aiohttp==3.7.4.post0
aiofiles==0.7.0
pymongo==3.12.0
ffmpeg==1.4.0
discord-py-slash-command==3.0.1
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
discord.py==1.7.2
dnspython==2.1.0
google==3.0.0
giphy-client==1.0.0
idna==2.10
instaloader==4.8.1
multidict==5.1.0
psutil==5.8.0
PyNaCl==1.4.0
pytz==2021.1
radio==0.1.3
progress==1.6
python-dotenv==0.17.1
requests==2.25.1
spotipy==2.19.0
typing-extensions==3.10.0.0
urllib3==1.26.5
wincertstore==0.2
yarl==1.6.3
youtube-dl==2021.6.6
zipp==3.4.1

